接口提交

+++ wydk/src/main/java/com/cc/wydk/entity/User.java
@@ -46,7 +46,7 @@ public class User implements UserDetails, Serializable {
 
     @ApiModelProperty("身份证")
     @TableField(value = "id_number")
-    private Integer idNumber;
+    private Long idNumber;
 
     @ApiModelProperty("政治面貌")
     @TableField(value = "political_outlook")
